requirements.txt
networkx
inflection
gunicorn==19.9
falcon
importlib_resources
PyYAML>=4.2b
setuptools
numpy
python-dateutil
profilehooks

# Zyre
pyzmq==17.1.2

# Mongo
pymodm
pymongo

# ROS
empy
rospkg
catkin-pkg
git+https://github.com:/ropod-project/rospy_message_converter.git

# ROPOD
git+https://github.com/ropod-project/fmlib.git@develop
git+https://github.com/ropod-project/mrta.git@develop#egg=mrs
git+https://github.com/ropod-project/mrta_stn.git@develop#egg=stn
git+https://github.com/ropod-project/osm_bridge.git

# Testing
pylint
pytest
pytest-cov
